refactor: Nest config, exception, and util package under common package
1. config, exception, util패키지를 common package하위로 이동시켰습니다.
