+++ src/main/java/com/fastshoppers/entity/BaseEntity.java
@@ -6,7 +6,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import com.fastshoppers.util.BooleanToCharConverter;
+import com.fastshoppers.common.util.BooleanToCharConverter;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
